Add files via upload
This is an example of call-back hell. or we can say basics of call back hell. why do we use it and what are its drawbacks?
In this simple example we make a pull request to the website dog API (https://dog.ceo/dog-api/documentation/breed ) and in return gives us the image link of a certain dog breed we call from the dog.txt and save it in another file dog.txt